requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
filelock==3.14.0
    # via
    #   huggingface-hub
    #   transformers
fsspec==2024.3.1
    # via huggingface-hub
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
huggingface-hub==0.23.0
    # via
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
itsdangerous==2.0.1
    # via -r requirements.in
jinja2==3.1.4
    # via fastapi
joblib==1.4.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   scikit-learn
    #   scipy
    #   transformers
orjson==3.10.3
    # via fastapi
packaging==24.0
    # via
    #   huggingface-hub
    #   transformers
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.4.28
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
rich==13.7.1
    # via typer
rsa==4.9
    # via python-jose
safetensors==0.4.3
    # via transformers
scikit-learn==1.4.2
    # via -r requirements.in
scipy==1.13.0
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   transformers
transformers==4.40.2
    # via -r requirements.in
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typer
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
